[BUGFIX] Avoid creating flash messages in CLI mode
TYPO3 FlashMessage can be stored to be persisted
accross requests, which is often a good thing,
but not in CLI environment.

`deepltranslate-glossary` dispatches FlashMessages
as part of a `DataHandler` hook implementation not
respecting that this must not be neccessary a web
request context and the absent of a session storage
leads to errors.

This change checks the environment now before adding
the dispatched FlashMessages without a fallback due
to the lack of a proper API to do so out of a hook
for the `DataHandler` in general CLI mode.

A try to mitigate this in TYPO3 directly has been
rejected [1] and states that extensions needs to
take care if this by themself, which is now done
with this change.

[1] https://review.typo3.org/c/Packages/TYPO3.CMS/+/91420/comments/39af877a_e6b81bb3